refactor the relayer
