+++ src/aptos/index.ts
@@ -1,3 +1,2 @@
 export * from './AptosNetwork';
 export * from './aptosNetworkUtils';
-export * from './serializer';
